Components: Add minimap button for addon compartment
The addon compartment was added in Dragonflight as an alternative way
to expose available addons to users, similar to what addons like
MinimapButton have provided.

This change puts the addon compartment into the same place as the
(optional) MBB button and ensures the associated dropdown menu opens
to the top and left of the button (because it itself is placed in the
bottom right).

If MinimapButton is installed, it will automatically hide the addon
compartment button to make place for its own alternative button.